backend/requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.6.0
PyJWT==2.10.1
PyMySQL==1.1.1
sqlacodegen==2.3.0.post1
SQLAlchemy==2.0.37
typeguard==4.4.1
typing_extensions==4.12.2
Werkzeug==3.1.3
